// src/core/enums.ts
export enum EntityType {
  CHARACTER = 'CHARACTER',
  MONSTER = 'MONSTER',
  NPC = 'NPC'
}

export enum ClassType {
  WARRIOR = 'WARRIOR',
  MAGE = 'MAGE',
  ROGUE = 'ROGUE',
  CLERIC = 'CLERIC'
}

export enum SkillType {
  ATTACK = 'ATTACK',
  HEAL = 'HEAL',
  BUFF = 'BUFF',
  DEBUFF = 'DEBUFF'
}

export enum ItemType {
  EQUIPMENT = 'EQUIPMENT',
  CONSUMABLE = 'CONSUMABLE',
  MATERIAL = 'MATERIAL'
}

export enum EquipmentSlot {
  HEAD = 'HEAD',
  BODY = 'BODY',
  HANDS = 'HANDS',
  FEET = 'FEET',
  WEAPON = 'WEAPON',
  OFFHAND = 'OFFHAND',
  ACCESSORY = 'ACCESSORY'
}

export enum AreaType {
  TOWN = 'TOWN',
  FOREST = 'FOREST',
  CAVE = 'CAVE',
  MOUNTAIN = 'MOUNTAIN',
  DUNGEON = 'DUNGEON'
}

export enum QuestType {
  MAIN = 'MAIN',
  SIDE = 'SIDE',
  DAILY = 'DAILY'
}

export enum ObjectiveType {
  KILL = 'KILL',
  COLLECT = 'COLLECT',
  EXPLORE = 'EXPLORE',
  TALK = 'TALK'
}